LayoutTests/storage/indexeddb/constants-expected.txt

Test IndexedDB's constants.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

PASS IDBDatabaseException.UNKNOWN_ERR is 1
PASS IDBDatabaseException.NON_TRANSIENT_ERR is 2
PASS IDBDatabaseException.CONSTRAINT_ERR is 4
PASS IDBDatabaseException.DATA_ERR is 5
PASS IDBDatabaseException.NOT_ALLOWED_ERR is 6
PASS IDBDatabaseException.TRANSACTION_INACTIVE_ERR is 7
PASS IDBDatabaseException.READ_ONLY_ERR is 9
PASS IDBDatabaseException.VER_ERR is 12
PASS IDBDatabaseException.TIMEOUT_ERR is DOMException.TIMEOUT_ERR
PASS IDBDatabaseException.QUOTA_ERR is DOMException.QUOTA_EXCEEDED_ERR
PASS IDBDatabaseException.ABORT_ERR is DOMException.ABORT_ERR
PASS IDBDatabaseException.NOT_FOUND_ERR is DOMException.NOT_FOUND_ERR
PASS successfullyParsed is true

TEST COMPLETE
